Remove the special note for "Selenium2-SAP". We do provide the pre-select verification in our implementation.
---  comments from Carl ----
If we have not provided pre-select verification--that is, verified the
item is present BEFORE we select it, then this note is valid.

This command does NOT normally provide any post-select verification
since Menu items are typically transient and gone once selected. So the
note is not necessary for the post-select case. Nowhere in the
description of this keyword do we suggest we do any post-select
verification.
-----------------------------